- 오토스캔 Page 작성
- 오토스캔 기능 구현중

+++ CalibrationNewGUI/Model/OthersInfo.cs
@@ -8,10 +8,10 @@ namespace CalibrationNewGUI.Model
 {
     public class OthersInfo
     {
-        public long InputVoltMax { get; set; } //입력 전압 범위(mV) 최대값
-        public long InputVoltMin { get; set; } //입력 전압 범위(mV) 최소값
-        public long InputCurrMax { get; set; } //입력 전류 범위(mA) 최대값
-        public long InputCurrMin { get; set; } //입력 전류 범위(mA) 최소값
+        public int InputVoltMax { get; set; } //입력 전압 범위(mV) 최대값
+        public int InputVoltMin { get; set; } //입력 전압 범위(mV) 최소값
+        public int InputCurrMax { get; set; } //입력 전류 범위(mA) 최대값
+        public int InputCurrMin { get; set; } //입력 전류 범위(mA) 최소값
         public string Language { get; set; } //현재 언어 설정값
 
         ConfigFileSave ConfigFile = new ConfigFileSave();
